feat: resolved review comments
